# Conflicts:
#	src/main/java/com/tcdt/qlnvhang/service/xuathang/xuatcuutrovientroxuatcap/xuatcuutrovientro/XhCtvtBbTinhKhoService.java

+++ src/main/java/com/tcdt/qlnvhang/request/xuathang/xuatcuutrovientroxuatcap/xuatcuutrovientro/XhCtvtBangKeReq.java
@@ -5,6 +5,7 @@ import com.tcdt.qlnvhang.request.BaseRequest;
 import com.tcdt.qlnvhang.table.xuathang.xuatcuutrovientroxuatcap.xuatcuutrovientro.XhCtvtBangKeDtl;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,9 +36,9 @@ public class XhCtvtBangKeReq extends BaseRequest {
   private String nlqDonVi;
   private String nlqDiaChi;
   private LocalDate thoiGianGiaoNhan;
-  private Long tongTrongLuong;
-  private Long tongTrongLuongBaoBi;
-  private Long tongTrongLuongHang;
+  private BigDecimal tongTrongLuong;
+  private BigDecimal tongTrongLuongBaoBi;
+  private BigDecimal tongTrongLuongHang;
   private String tongTrongLuongHangBc;
   private LocalDate ngayGduyet;
   private Long nguoiGduyetId;
